Compare found hash vs. hash-object output
This updates the find function to also return the hash it found, in
preparation to add a flag to make writing the commit object optional.

The found hash should never differ from what hash-object returns, but
since we have it now we might as well add an extra check just in case
something changes in the future.